Separate reduction action from API call
